code changes for leave request

+++ HRConnect.Application/DTO/Leave/LeaveDashboardDto.cs
@@ -8,20 +8,20 @@ namespace HRConnect.Application.DTO.Leave
 {
     public class LeaveDashboardDto
     {
-        public int CasualLeaveTotal { get; set; }
-        public int CasualLeaveUsed { get; set; }
-        public int CasualLeaveRemaining { get; set; }
+        public decimal CasualLeaveTotal { get; set; }
+        public decimal CasualLeaveUsed { get; set; }
+        public decimal CasualLeaveRemaining { get; set; }
 
-        public int SickLeaveTotal { get; set; }
-        public int SickLeaveUsed { get; set; }
-        public int SickLeaveRemaining { get; set; }
+        public decimal SickLeaveTotal { get; set; }
+        public decimal SickLeaveUsed { get; set; }
+        public decimal SickLeaveRemaining { get; set; }
 
-        public int EarnedLeaveTotal { get; set; }
-        public int EarnedLeaveUsed { get; set; }
-        public int EarnedLeaveRemaining { get; set; }
+        public decimal EarnedLeaveTotal { get; set; }
+        public decimal EarnedLeaveUsed { get; set; }
+        public decimal EarnedLeaveRemaining { get; set; }
 
-        public int PendingRequests { get; set; }
-        public int ApprovedRequests { get; set; }
-        public int RejectedRequests { get; set; }
+        public decimal PendingRequests { get; set; }
+        public decimal ApprovedRequests { get; set; }
+        public decimal RejectedRequests { get; set; }
     }
 }
